add dump_png util
