+++ cpp/apps/CMakeLists.txt
@@ -10,10 +10,7 @@ if(RESIZE_PNG)
   add_subdirectory(resize_png)
 endif(RESIZE_PNG)
 
-# option(AVIF_DUMP "Build the avif_dump binary" ON) if(AVIF_DUMP)
-# add_subdirectory(avif_dump) endif(AVIF_DUMP)
-
-option(WEBP_DUMP "Build the webp_dump binary" ON)
-if(WEBP_DUMP)
-  add_subdirectory(webp_dump)
-endif(WEBP_DUMP)
+option(DUMP_PNG "Build the dump_png binary" ON)
+if(DUMP_PNG)
+  add_subdirectory(dump_png)
+endif(DUMP_PNG)
